Changed move encoding

+++ include/chesslogic.h
@@ -4,7 +4,7 @@
 
 class ChessLogic {
 private:
-    std::map<u32, std::string> fens;
+    std::map<u64, std::string> fens;
     Board board;
     bool white_turn;
     bool wk_castle, wq_castle, bk_castle, bq_castle;
@@ -12,14 +12,14 @@ private:
     int en_passant_square;
     int half_moves;
     int full_moves;
-    u32 total_moves;
+    u64 total_moves;
 
 public:
     ChessLogic();
     void load_pos(const std::string &);
     void draw_game();
-    bool make_move(u32);
-    bool unmake_move(u32);
+    bool make_move(u64);
+    bool unmake_move(u64);
     Board &get_board();
     void change_turn();
     void set_turn(bool);
@@ -44,8 +44,8 @@ public:
     void set_doublecheck(bool);
     bool fivefold_repitition();
     bool fifty_move_rule();
-    u32 get_totalmoves();
-    std::map<u32, std::string> get_fens();
+    u64 get_totalmoves();
+    std::map<u64, std::string> get_fens();
 };
 
 #endif // !CHESSLOGIC_H
